+++ src/index.ts
@@ -3,4 +3,9 @@ import './locales/index.js'
 export * from './hooks'
 export * as constants from './constants'
 export * from './period-calculation'
-export { default as getNowInCalendar } from './utils/getNowInCalendar'
+export {
+    getNowInCalendar,
+    validateDateString,
+    convertFromIso8601,
+    convertToIso8601,
+} from './utils'