+++ src/period-calculation/yearly-periods/build-yearly-fixed-period.ts
@@ -1,10 +1,6 @@
 import { Temporal } from '@js-temporal/polyfill'
 import { SupportedCalendar } from '../../types'
-import {
-    fromAnyDate,
-    formatYyyyMmDD,
-    isCustomCalendar,
-} from '../../utils/index'
+import { fromAnyDate, formatDate, isCustomCalendar } from '../../utils/index'
 import localisationHelpers from '../../utils/localisationHelpers'
 import { financialYearFixedPeriodTypes } from '../period-type-groups'
 import { FixedPeriod, PeriodType } from '../types'
@@ -47,8 +43,8 @@ const buildYearlyFixedPeriod: BuildYearlyFixedPeriod = ({
         iso: value,
         name,
         displayName: name,
-        startDate: formatYyyyMmDD(startDate, 'startOfMonth'),
-        endDate: formatYyyyMmDD(endDate, 'endOfMonth'),
+        startDate: formatDate(startDate, 'startOfMonth'),
+        endDate: formatDate(endDate, 'endOfMonth'),
     }
 }