+++ src/utils/extract-date-parts-from-date-string.ts
@@ -1,0 +1,48 @@
+import i18n from '@dhis2/d2-i18n'
+import { dateStringRegExp } from './date-string-regexp'
+
+export function extractDatePartsFromDateString(
+    dateString: string,
+    format?: 'YYYY-MM-DD' | 'DD-MM-YYYY'
+) {
+    const parts = dateString.match(dateStringRegExp)
+
+    if (!parts) {
+        throw new Error(
+            i18n.t(`Date string is invalid, received "{{dateString}}"`, {
+                dateString,
+            })
+        )
+    }
+
+    let yearStr, monthStr, dayStr, detectedFormat
+
+    if (parts[1]) {
+        // Match for YYYY-MM-DD
+        yearStr = parts[1]
+        monthStr = parts[3]
+        dayStr = parts[5]
+        detectedFormat = 'YYYY-MM-DD'
+    } else {
+        // Match for DD-MM-YYYY
+        dayStr = parts[6]
+        monthStr = parts[8]
+        yearStr = parts[10]
+        detectedFormat = 'DD-MM-YYYY'
+    }
+
+    if (format && detectedFormat !== format.toUpperCase()) {
+        throw new Error(
+            i18n.t(
+                `Date string format does not match the specified format. Expected {{format}} but got {{detectedFormat}}.`,
+                { format, detectedFormat }
+            )
+        )
+    }
+
+    const year = parseInt(yearStr, 10)
+    const month = parseInt(monthStr, 10)
+    const day = parseInt(dayStr, 10)
+
+    return { year, month, day, format: detectedFormat }
+}