+++ src/utils/index.ts
@@ -1,4 +1,8 @@
+export { extractDatePartsFromDateString } from './extract-date-parts-from-date-string'
 export { default as fromAnyDate } from './from-any-date'
 export { default as getNowInCalendar } from './getNowInCalendar'
 export * from './helpers'
 export { default as localisationHelpers } from './localisationHelpers'
+export { validateDateString } from './validate-date-string'
+
+export { convertFromIso8601, convertToIso8601 } from './convert-date'